test: improve test coverage for Toast and ToastManager
